refactor: remove liquidity pool and staking farm
