+++ contracts/linear/src/lib.rs
@@ -1,6 +1,6 @@
 use near_sdk::{
     borsh::{self, BorshDeserialize, BorshSerialize},
-    collections::{UnorderedMap, UnorderedSet, Vector},
+    collections::{UnorderedMap, UnorderedSet},
     env, ext_contract,
     json_types::U128,
     near_bindgen, require,
@@ -12,11 +12,9 @@ mod account;
 mod epoch_actions;
 mod errors;
 mod events;
-mod farm;
 mod fungible_token;
 mod internal;
 mod legacy;
-mod liquidity_pool;
 mod metadata;
 mod owner;
 mod stake;
@@ -28,9 +26,7 @@ mod view;
 
 use crate::account::*;
 use crate::errors::*;
-use crate::farm::Farm;
 use crate::fungible_token::*;
-use crate::liquidity_pool::*;
 use crate::types::*;
 use crate::utils::*;
 use crate::validator_pool::*;
@@ -42,9 +38,7 @@ pub(crate) enum StorageKey {
     Accounts,
     Shares,
     Beneficiaries,
-    Validators,           // V0 (Don't comment out this)
-    Farms,                // [DEPRECATED]
-    AuthorizedFarmTokens, // [DEPRECATED]
+    Validators, // V0 (Don't comment out this)
     Managers,
     ValidatorsV1, // Used in v1.3.0 upgrade
 }
@@ -79,9 +73,6 @@ pub struct LiquidStakingContract {
     /// Beneficiaries for staking rewards.
     beneficiaries: UnorderedMap<AccountId, u32>,
 
-    /// [DEPRECATED] The single-direction liquidity pool that enables instant unstake
-    liquidity_pool: LiquidityPool,
-
     // --- Validator Pool ---
     /// The validator pool that manage the actions against validators
     validator_pool: ValidatorPool,
@@ -98,15 +89,6 @@ pub struct LiquidStakingContract {
     unstake_amount_to_settle: Balance,
     /// Last epoch height stake/unstake settlements were calculated
     last_settlement_epoch: EpochHeight,
-
-    // --- [DEPRECATED] Staking Farm ---
-    /// [DEPRECATED] Farm tokens.
-    farms: Vector<Farm>,
-    /// [DEPRECATED] Active farms: indicies into `farms`.
-    active_farms: Vec<u64>,
-    /// [DEPRECATED] Authorized tokens for farms.
-    /// Required because any contract can call method with ft_transfer_call, so must verify that contract will accept it.
-    authorized_farm_tokens: UnorderedSet<AccountId>,
 }
 
 #[near_bindgen]
@@ -145,7 +127,6 @@ impl LiquidStakingContract {
             paused: false,
             account_storage_usage: 0,
             beneficiaries: UnorderedMap::new(StorageKey::Beneficiaries),
-            liquidity_pool: LiquidityPool::new(LiquidityPoolConfig::default()),
             // Validator Pool
             validator_pool: ValidatorPool::new(),
             whitelist_account_id: None,
@@ -154,10 +135,6 @@ impl LiquidStakingContract {
             stake_amount_to_settle: 0,
             unstake_amount_to_settle: 0,
             last_settlement_epoch: 0,
-            // [DEPRECATED] Staking Farm
-            farms: Vector::new(StorageKey::Farms),
-            active_farms: Vec::new(),
-            authorized_farm_tokens: UnorderedSet::new(StorageKey::AuthorizedFarmTokens),
         };
         this.internal_add_manager(&owner_id);
         this.measure_account_storage_usage();
